refactor(hooks): rename slice files to kebab-case and update imports
