+++ packages/giselle/src/react/flow/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from "../utils/is-supported-connection";
 
-export * from "./fileSlice";
-export * from "./propertiesPanelSlice";
+export * from "./file-slice";
+export * from "./properties-panel-slice";
 export * from "./store";
 export * from "./use-add-node";
 export * from "./use-connection";
@@ -10,5 +10,5 @@ export * from "./use-node-update";
 export * from "./use-properties-panel";
 export * from "./use-workflow-designer";
 export * from "./use-workspace-reducer";
-export * from "./viewSlice";
-export * from "./workspaceSlice";
+export * from "./view-slice";
+export * from "./workspace-slice";
