+++ packages/giselle/src/react/flow/hooks/properties-panel-slice.ts
