+++ packages/giselle/src/react/flow/hooks/store.ts
@@ -1,11 +1,11 @@
 import { create } from "zustand";
-import { createFileSlice, type FileSlice } from "./fileSlice";
+import { createFileSlice, type FileSlice } from "./file-slice";
 import {
 	createPropertiesPanelSlice,
 	type PropertiesPanelSlice,
-} from "./propertiesPanelSlice";
-import { createViewSlice, type ViewSlice } from "./viewSlice";
-import { createWorkspaceSlice, type WorkspaceSlice } from "./workspaceSlice";
+} from "./properties-panel-slice";
+import { createViewSlice, type ViewSlice } from "./view-slice";
+import { createWorkspaceSlice, type WorkspaceSlice } from "./workspace-slice";
 
 // The full store type is the intersection of all slice types
 export type AppStore = WorkspaceSlice &
