+++ packages/giselle/src/react/flow/index.ts
@@ -4,4 +4,4 @@ export * from "./context";
 export * from "./hooks";
 export * from "./types";
 export * from "./utils";
-export * from "./ZustandBridgeProvider";
+export * from "./zustand-bridge-provider";
